backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-filter==24.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
importlib_metadata==8.5.0
Markdown==3.7
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
zipp==3.20.2
